v9.0

+++ manage_users_controller.php
@@ -5,6 +5,12 @@ require_once('Models/User.php');
 $db = new connectionDB();
 $conn = $db->connect();
 $userModel = new User($conn);
+session_start();
+$is_admin = ($_SESSION['user_role'] === "Admin");
+if (!isset($_SESSION['user_id']) || !isset($_SESSION['user_role']) ||  !$is_admin) {
+    header("Location: /Borrow_My_Charger/login.php");
+    exit;
+}
 
 if (isset($_GET['delete'])) {
     $userModel->deleteUser($_GET['delete']);
